evals/requirements.txt
openai
numpy
pandas
pytest
python-dotenv
